fix: comprehensive test infrastructure overhaul
- Convert integer foreign keys to bigint for Rails 8.0+ compatibility
- Fix PostgreSQL transaction abort issues with proper destroy_all teardown
- Remove MySQL-incompatible timestamp defaults from migrations
- Update Makefile with proper error propagation for CI failure detection
- Ensure schema consistency across all database adapters

Resolves foreign key constraint violations and column type mismatches
that were causing hidden test failures in CI.

+++ test/dummy/db/schema.rb
@@ -12,9 +12,9 @@
 
 ActiveRecord::Schema[7.2].define(version: 17) do
   create_table "application_taggings", force: :cascade do |t|
-    t.integer "tag_id", null: false
+    t.bigint "tag_id", null: false
     t.string "taggable_type", null: false
-    t.integer "taggable_id", null: false
+    t.bigint "taggable_id", null: false
     t.string "context", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
